feat: add module createPage
